Corrected potential PROGMEM const variables access issue
Most PROGMEM  const variables needs to be in lower 64K  of flash as accessing most of them assumes this.  So added a PROGMEMSECTION2 to move those that are always accessed as far addresses (can be anywhere in the address space) into that new region to avoid having PROGMEM  const variables overflow beyond lower 64K regsion So moved the Dayo, and WereWolves of London setups into PROGMEMSECTION2 areas are assumed to be anywhere in flash memory and accessed via far routines.