Corrected servo angle limit
PWMServoNew angle member variable was uint8_t and thus limited angle to 0 - 256 .  Needed angle to go to at least 300 so changed it to uint16_t.  Also replaced min/max16 which are 16usec units with minPulseWidth/maxPulseWidth in usec and removed the * 16L in write to reduce confusion (i.e. why divide by 16 and then turn around and multiply by 16 in write if that is the only place min/max16 are used)