Moded ImuHeadMouth  cmd to add live synchronized Arm, Elbow movements.
Moded ImuHeadMouth  cmd to add Live Arm, Elbow movements.   Tested using  arm IMUs hinge block by moving it by hand (not strapped to arm yet, but will try that next).  Pretty cool and mostly works with a few issues 1)  When arm IMU roll gets over 40 degrees or so the prop elbow which was bent at 90 degrees on the IMU hinge block appears to straighten as the roll increases for some reason even though  I am not straightening the elbow on the IMU hinge block.  2)  Sometimes the prop arm will briefly stop and then jump to the final position  when moving arm IMU block swiftly.  Not sure if IMU position read is causing this by going out of range for a second or what.