Updates for getting mimicing movement with arm brace mostly working
Code Updates for getting prop synchronized  movement to arm brace IMUs mostly working.  There are still weird transitions that sometimes occur where the prop arm goes to some wild position momentarily or statically  when the elbow / arm imu pitch is more than 133 degrees (so limit IMU usage to no more than that but maybe don't even update Yaw in that case in the IMU filter?)..  So it's mostly an IMU issue with certain positions.  like also 90 degrees pitch, can also cause weird Yaw values too with certain motions.  Tried using quaternions and then converted back to Euler angles for final use with servo but it didn't make a difference as the final IMU angles still had the same issue.  Possibly need to look at the  quaternion to Euler routine to see if it has an issue.   Also , could possibly use two IMUs each on the arm and elbow and with one IMU offset by an angle to the other and then use the reading from the one that is not at the weird pitch angle.  The fact that the servos are setup as yaw, roll, pitch configuration makes things difficult as well as would need to recognize the transitions where one of those jumps dramatically, but still in range, due to singularities and some how account for that