Mods for reordering arm servos
Changed  hardware order of servos on arm to match normal Euler angle order and what the IMU sensors assume for yaw, pitch, roll.   So now the pitch servo is connected to yaw shaft , and roll servo is connected to pitch shaft whereas before the order was roll servo connected to yaw shaft and pitch servo connected to roll shaft.  The order is important as a given yaw, pitch, roll value does not yield the same arm orientation with a different servo order (i.e. which shaft the yaw, pitch, roll servos are attached to).  Once did this the prop arm now seems to reasonably follow/mimic the  user arm movements, although more improvements with the arm brace and sensor placement on the user  can still be made to mitigate any false detected movements or orientation.   Updated description in the readme.md file and added new pics to reflect the above change.  Also a minor mod to the PropController.into file to limit the max arm Yaw position to keep teh servos from hitting each other.